Added authentication mechanism.
A BRemote can perform authentication internally if the session is
expired or invalid. 
If BRemote is tagged with @authparam <paramclassname>, the parameter of
this class name is assumed to be a session object that passes the
authentication to the server. The BStub methods are generated without
this parameter. Internally, the parameter is supplied from
BAuthentication.getSessionObject().